+++ RELEASE_NOTES.md
@@ -1,6 +1,7 @@
 #### 1.3.5 - Unreleased
 
 * Fix null reference exception when request definition is merged and doesn't have explicit element (and content name).
+* Restore `Null` option of empty tag handling (provides different behavior for nullable value types).
 
 #### 1.3.4 - 07.12.2018